cooking-ai-agent/requirements.txt
# Azure AI Inference SDK for GitHub Models
azure-ai-inference>=1.0.0b1
azure-core>=1.29.0

# Optional: for enhanced JSON handling
requests>=2.31.0
